Add proper descriptions for inspections - Optimize package structure for dependency inspection - Code cleanup
