+++ src/main/kotlin/me/schlaubi/intellij_gradle_version_checker/ProjectGradleVersionCheckingActivity.kt
@@ -26,11 +26,8 @@ package me.schlaubi.intellij_gradle_version_checker
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vcs.changes.shelf.ShelvedChangesViewManager
-import com.jetbrains.rd.util.first
 import me.schlaubi.intellij_gradle_version_checker.settings.ApplicationGradleVersionSettings
 import me.schlaubi.intellij_gradle_version_checker.settings.ProjectPersistentGradleVersionSettings
-import org.jetbrains.plugins.gradle.settings.GradleExtensionsSettings
-import org.jetbrains.plugins.gradle.settings.GradleLocalSettings
 import org.jetbrains.plugins.gradle.settings.GradleSettings
 import org.gradle.util.GradleVersion as GGradleVersion
