+++ src/main/resources/inspectionDescriptions/OutdatedGradleVersion.html
@@ -1,0 +1,35 @@
+<!--
+  ~ MIT License
+  ~
+  ~ Copyright (c) 2020 Michael Rittmeister
+  ~
+  ~ Permission is hereby granted, free of charge, to any person obtaining a copy
+  ~ of this software and associated documentation files (the "Software"), to deal
+  ~ in the Software without restriction, including without limitation the rights
+  ~ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+  ~ copies of the Software, and to permit persons to whom the Software is
+  ~ furnished to do so, subject to the following conditions:
+  ~
+  ~ The above copyright notice and this permission notice shall be included in all
+  ~ copies or substantial portions of the Software.
+  ~
+  ~ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+  ~ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+  ~ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+  ~ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+  ~ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+  ~ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+  ~ SOFTWARE.
+  -->
+
+<html lang="en">
+<body>
+<p>
+  This Inspection checks whether you use the latest Gradle Version or not. It will also have various severities
+  depending on how old your Gradle version is
+</p>
+<!-- tooltip end -->
+Unfortunately IntelliJ IDEA often uses an outdated Gradle Version when creating a brand new project. Check
+<a href="https://youtrack.jetbrains.com/issue/IDEA-215792">IDEA-215792</a> for more information.
+</body>
+</html>
