+++ src/main/resources/messages/GradleUpdater.properties
@@ -23,8 +23,11 @@
 #
 inspection.group.name=Gradle version inspections
 inspection.outdated_version.description=Gradle version is outdated! Use {0}.
-inspection.outdated_version.static_description=This inspection checks the version of you gradle-wrapper and compares it to the latest Gradle release
-inspection.deprecated_dependency_config.description=The 'compile' dependency configuration got deprecated and will be removed in Gradle 6. Therefore you should replace it with the 'implementation' dependency config
+inspection.deprecated_dependency_config.description=The 'compile' dependency configuration got deprecated and will be removed in a newer Gradle version. Therefore you should replace it with the 'implementation' dependency config
+inspection.kotlin_plugin_with_id.description=Not idiomatic plugin declaration of Kotlin Gradle Plugin
+inspection.kotlin_plugin_with_id.quickfix.replace=Replace with kotlin() notation
+inspection.built_in_plugin_with_id.description=Not idiomatic plugin declaration of built-in Gradle Plugin
+inspection.built_in_plugin_with_id.quickfix.replace=Replace with literal notation
 quickfix.update_gradle.family_name=Update to latest Gradle version
 quickfix.update_gradle_and_sync.family_name=Update to latest Gradle version and sync Gradle project
 quickfix.migrate_to_implementation.family_name=Replace with 'implementation'
@@ -41,5 +44,8 @@ notification.outdated_version.actions.ignore_project.description=Mutes this warn
 configurable.group.tools.gradle_updater.settings.display.name=Gradle Version Updater Settings
 configurable.group.tools.gradle_updater.settings.display.description=<html><body>\
   Configure your project integration with Android build tools.
+# suppress inspection "SpellCheckingInspection"
 copypaste.onpaste.seems.like.groovy=Pasted Code Looks Like Groovy
+# suppress inspection "SpellCheckingInspection"
 copypaste.onpaste.seems.like.groovy.description=Clipboard content looks like Groovy DSL dependency. Do you want to convert it to Kotlin DSL?
+
